requirements.txt
yfinance
pyzmail36
imapclient
python-telegram-bot==13.15
python-dotenv
